refactor(discord): bootstrap の Prometheus メトリクス生成を metrics.ts へ分離する
