+++ apps/discord/src/bootstrap.ts
@@ -33,7 +33,7 @@ import {
 } from "@vicissitude/memory/namespace";
 import { MemoryStorage } from "@vicissitude/memory/storage";
 import { ConsoleLogger } from "@vicissitude/observability/logger";
-import { PrometheusCollector, PrometheusServer, METRIC } from "@vicissitude/observability/metrics";
+import { METRIC, type PrometheusCollector } from "@vicissitude/observability/metrics";
 import { OllamaEmbeddingAdapter } from "@vicissitude/ollama";
 import {
 	denyAllSkillPermission,
@@ -67,6 +67,7 @@ import {
 	createStoreLayer,
 	createWebContextLayer,
 } from "./bootstrap/layers.ts";
+import { createMetrics } from "./bootstrap/metrics.ts";
 import { type AppConfig, loadConfig } from "./config.ts";
 import { ChannelConfigLoader, type ChannelConfigData } from "./gateway/channel-config-loader.ts";
 import { DiscordGateway } from "./gateway/discord.ts";
@@ -79,48 +80,6 @@ import {
 import { MoodNicknameService } from "./mood-nickname.ts";
 import { createPortLayout } from "./port-allocator.ts";
 import { createShutdown } from "./shutdown.ts";
-
-// ─── Metrics ────────────────────────────────────────────────────
-
-export function createMetrics(logger: Logger, port: number) {
-	const collector = new PrometheusCollector();
-	collector.registerCounter(METRIC.DISCORD_MESSAGES_RECEIVED, "Discord messages received");
-	collector.registerCounter(METRIC.AI_REQUESTS, "Completed AI prompt requests");
-	collector.registerCounter(METRIC.HEARTBEAT_TICKS, "Heartbeat scheduler ticks");
-	collector.registerCounter(METRIC.HEARTBEAT_REMINDERS_EXECUTED, "Heartbeat reminders executed");
-	collector.registerGauge(METRIC.BOT_INFO, "Bot information");
-	collector.registerHistogram(METRIC.AI_REQUEST_DURATION, "AI prompt duration in seconds");
-	collector.registerHistogram(METRIC.HEARTBEAT_TICK_DURATION, "Heartbeat tick duration in seconds");
-	collector.registerGauge(METRIC.LLM_ACTIVE_SESSIONS, "Registered LLM sessions");
-	collector.registerGauge(METRIC.LLM_BUSY_SESSIONS, "LLM prompts currently processing");
-	collector.registerCounter(
-		METRIC.MEMORY_CONSOLIDATION_TICKS,
-		"Memory consolidation scheduler ticks",
-	);
-	collector.registerHistogram(
-		METRIC.MEMORY_CONSOLIDATION_TICK_DURATION,
-		"Memory consolidation tick duration in seconds",
-	);
-	// Token metrics
-	collector.registerCounter(METRIC.LLM_INPUT_TOKENS, "LLM input tokens total");
-	collector.registerCounter(METRIC.LLM_OUTPUT_TOKENS, "LLM output tokens total");
-	collector.registerCounter(METRIC.LLM_CACHE_READ_TOKENS, "LLM cache read tokens total");
-	// Cost metrics
-	collector.registerCounter(METRIC.LLM_COST_DOLLARS, "LLM cost in US dollars");
-	// Session error metrics
-	collector.registerCounter(METRIC.SESSION_ERRORS, "Session errors total");
-	collector.registerCounter(METRIC.SESSION_RESTARTS, "Session restarts total");
-	collector.registerCounter(METRIC.SESSION_RETRIES, "Session retries total");
-	// Emotion estimation metrics
-	collector.registerCounter(METRIC.EMOTION_ESTIMATION_ERRORS, "Emotion estimation errors total");
-	collector.registerCounter(METRIC.EMOTION_ESTIMATION_SKIPS, "Emotion estimation skips total");
-	// Drift metrics
-	collector.registerGauge(METRIC.DRIFT_SCORE, "Character drift score per guild");
-	collector.registerCounter(METRIC.DRIFT_AUDITS, "Character drift audit results");
-	collector.registerCounter(METRIC.CRITIC_AUDITOR_SKIP_TOTAL, "Critic auditor skipped audits");
-	collector.setGauge(METRIC.BOT_INFO, 1, { bot_name: "hua" });
-	return { collector, server: new PrometheusServer(collector, logger, port) };
-}
 
 // ─── Channel Config ─────────────────────────────────────────────
